src/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
azure-identity==1.19.0
aiohttp==3.11.1
azure-ai-inference[prompts]
azure-ai-projects